+++ src/App.js
@@ -1,37 +1,15 @@
-import Categories from "./components/categories/categories.component"
+import {Routes, Route, Outlet} from 'react-router-dom'
 
-const categories = [
-  {
-    "id": 1,
-    "title": "hats",
-    "imageUrl": "https://i.ibb.co/cvpntL1/hats.png"
-  },
-  {
-    "id": 2,
-    "title": "jackets",
-    "imageUrl": "https://i.ibb.co/px2tCc3/jackets.png"
-  },
-  {
-    "id": 3,
-    "title": "sneakers",
-    "imageUrl": "https://i.ibb.co/0jqHpnp/sneakers.png"
-  },
-  {
-    "id": 4,
-    "title": "womens",
-    "imageUrl": "https://i.ibb.co/GCCdy8t/womens.png"
-  },
-  {
-    "id": 5,
-    "title": "mens",
-    "imageUrl": "https://i.ibb.co/R70vBrQ/men.png"
-  }
-]
+import Home from "./routes/home/home.component";
+import Navigation from './routes/navigation/navigation.component';
 
 const App = () => {
-
-  return (
-    <Categories categories={categories}/>
+  return ( 
+    <Routes>
+      <Route path='/' element={<Navigation/>}>
+        <Route index element={<Home/>}/>
+      </Route>
+    </Routes>
   );
 }
